Fix maximum constant size for D3D11/12

+++ Backends/Graphics4/Direct3D11/Sources/Kore/Direct3D11.winrt.cpp
@@ -453,26 +453,26 @@ void Graphics4::setTextureOperation(TextureOperation operation, TextureArgument 
 }
 
 namespace {
-	void setInt(u8* constants, u8 offset, u8 size, int value) {
+	void setInt(u8* constants, u32 offset, u32 size, int value) {
 		if (size == 0) return;
 		int* ints = reinterpret_cast<int*>(&constants[offset]);
 		ints[0] = value;
 	}
 
-	void setFloat(u8* constants, u8 offset, u8 size, float value) {
+	void setFloat(u8* constants, u32 offset, u32 size, float value) {
 		if (size == 0) return;
 		float* floats = reinterpret_cast<float*>(&constants[offset]);
 		floats[0] = value;
 	}
 
-	void setFloat2(u8* constants, u8 offset, u8 size, float value1, float value2) {
+	void setFloat2(u8* constants, u32 offset, u32 size, float value1, float value2) {
 		if (size == 0) return;
 		float* floats = reinterpret_cast<float*>(&constants[offset]);
 		floats[0] = value1;
 		floats[1] = value2;
 	}
 
-	void setFloat3(u8* constants, u8 offset, u8 size, float value1, float value2, float value3) {
+	void setFloat3(u8* constants, u32 offset, u32 size, float value1, float value2, float value3) {
 		if (size == 0) return;
 		float* floats = reinterpret_cast<float*>(&constants[offset]);
 		floats[0] = value1;
@@ -480,7 +480,7 @@ namespace {
 		floats[2] = value3;
 	}
 
-	void setFloat4(u8* constants, u8 offset, u8 size, float value1, float value2, float value3, float value4) {
+	void setFloat4(u8* constants, u32 offset, u32 size, float value1, float value2, float value3, float value4) {
 		if (size == 0) return;
 		float* floats = reinterpret_cast<float*>(&constants[offset]);
 		floats[0] = value1;
@@ -489,7 +489,7 @@ namespace {
 		floats[3] = value4;
 	}
 
-	void setFloats(u8* constants, u8 offset, u8 size, float* values, int count) {
+	void setFloats(u8* constants, u32 offset, u32 size, float* values, int count) {
 		if (size == 0) return;
 		float* floats = reinterpret_cast<float*>(&constants[offset]);
 		for (int i = 0; i < count; ++i) {
@@ -497,13 +497,13 @@ namespace {
 		}
 	}
 
-	void setBool(u8* constants, u8 offset, u8 size, bool value) {
+	void setBool(u8* constants, u32 offset, u32 size, bool value) {
 		if (size == 0) return;
 		int* ints = reinterpret_cast<int*>(&constants[offset]);
 		ints[0] = value ? 1 : 0;
 	}
 
-	void setMatrix(u8* constants, u8 offset, u8 size, const mat4& value) {
+	void setMatrix(u8* constants, u32 offset, u32 size, const mat4& value) {
 		if (size == 0) return;
 		float* floats = reinterpret_cast<float*>(&constants[offset]);
 		for (int y = 0; y < 4; ++y) {
@@ -513,7 +513,7 @@ namespace {
 		}
 	}
 
-	void setMatrix(u8* constants, u8 offset, u8 size, const mat3& value) {
+	void setMatrix(u8* constants, u32 offset, u32 size, const mat3& value) {
 		if (size == 0) return;
 		float* floats = reinterpret_cast<float*>(&constants[offset]);
 		for (int y = 0; y < 3; ++y) {
